pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "智能数据分析系统Demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "sqlalchemy>=1.4.0",
    "dashscope>=1.10.0",
    "python-dotenv>=0.19.0",
    "PyMySQL>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"